#59 Kompatibel mit Postgres machen

See merge request namib/mud-controller-enforcer/namib_mud_controller!40

+++ migrations/postgres/20210124172420_create_user_config.sql
@@ -1,9 +1,9 @@
 -- Your SQL goes here
 CREATE TABLE user_configs
 (
-    key         TEXT NOT NULL,
-    user_id     INTEGER NOT NULL,
-    value       TEXT NOT NULL DEFAULT '',
+    key     TEXT   NOT NULL,
+    user_id BIGINT NOT NULL,
+    value   TEXT   NOT NULL DEFAULT '',
     PRIMARY KEY (key, user_id),
     FOREIGN KEY (user_id) REFERENCES users (id)
         ON DELETE CASCADE ON UPDATE NO ACTION
